requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.3
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfminer.six==20231228